Consolidate character tests to a single class
